move dom contents into dom.js
